backend/requirements.txt
fastapi>=0.115.6
uvicorn[standard]>=0.32.1
python-dotenv>=1.0.1
slowapi>=0.1.9
pydantic>=2.11.0
agent-framework-core>=1.0.0rc4
